requirements.txt
certifi==2020.12.5
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
speedtest-cli==2.1.2
wincertstore==0.2
